Move Role(s) to `sc-network-common`
