+++ client/network/src/lib.rs
@@ -260,7 +260,8 @@ pub use libp2p::{multiaddr, Multiaddr, PeerId};
 pub use protocol::PeerInfo;
 pub use sc_network_common::{
 	protocol::{
-		event::{DhtEvent, Event, ObservedRole},
+		event::{DhtEvent, Event},
+		role::ObservedRole,
 		ProtocolName,
 	},
 	request_responses::{IfDisconnected, RequestFailure},
