+++ client/network/transactions/src/lib.rs
@@ -35,10 +35,7 @@ use prometheus_endpoint::{register, Counter, PrometheusError, Registry, U64};
 use sc_network_common::{
 	config::{NonDefaultSetConfig, NonReservedPeerMode, ProtocolId, SetConfig},
 	error,
-	protocol::{
-		event::{Event, ObservedRole},
-		ProtocolName,
-	},
+	protocol::{event::Event, role::ObservedRole, ProtocolName},
 	service::{NetworkEventStream, NetworkNotification, NetworkPeers},
 	utils::{interval, LruHashSet},
 	ExHashT,
